+++ src/main/java/com/doldolseo/doldolseo_msa_crew_board/controller/CrewPostController.java
@@ -1,12 +1,13 @@
 package com.doldolseo.doldolseo_msa_crew_board.controller;
 
-import com.doldolseo.doldolseo_msa_crew_board.dto.CrewPostAndMembersWithDTO;
+import com.doldolseo.doldolseo_msa_crew_board.domain.TaggedMemberId;
+import com.doldolseo.doldolseo_msa_crew_board.dto.CrewPostAndMembersDTO;
 import com.doldolseo.doldolseo_msa_crew_board.dto.CrewPostDTO;
 import com.doldolseo.doldolseo_msa_crew_board.dto.CrewPostPageDTO;
 import com.doldolseo.doldolseo_msa_crew_board.service.CrewPostService;
+import com.doldolseo.doldolseo_msa_crew_board.utils.AuthorityUtil;
 import com.doldolseo.doldolseo_msa_crew_board.utils.UploadCrewPostFileUtil;
 import org.apache.commons.io.IOUtils;
-import org.apache.tomcat.util.http.fileupload.impl.FileSizeLimitExceededException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
@@ -17,10 +18,10 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import javax.servlet.http.HttpServletRequest;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.Optional;
 
 @RestController
 public class CrewPostController {
@@ -28,6 +29,14 @@ public class CrewPostController {
     CrewPostService service;
     @Autowired
     UploadCrewPostFileUtil fileUtil;
+    @Autowired
+    AuthorityUtil authorityUtil;
+
+    @PostMapping(value = "/crew/post")
+    public ResponseEntity<CrewPostDTO> createPost(CrewPostDTO dtoIn, String taggedMembers) {
+        service.saveCrewPost(dtoIn, taggedMembers);
+        return ResponseEntity.status(HttpStatus.OK).body(null);
+    }
 
     @GetMapping(value = "/crew/post")
     public ResponseEntity<CrewPostPageDTO> getCrewPostPage(CrewPostDTO dtoIn,
@@ -36,31 +45,47 @@ public class CrewPostController {
     }
 
     @GetMapping(value = "/crew/post/{crewPostNo}")
-    public ResponseEntity<CrewPostAndMembersWithDTO> getPost(@PathVariable(value = "crewPostNo") Long crewPostNo,
-                                                             @RequestParam(required = false) String doHit) throws Exception {
-        if (doHit.equals("yes"))
-            return ResponseEntity.status(HttpStatus.OK).body(service.getCrewPostAndMembersWithAndHit(crewPostNo));
+    public ResponseEntity<CrewPostAndMembersDTO> getPost(@PathVariable(value = "crewPostNo") Long crewPostNo,
+                                                         @RequestParam(required = false) Optional<String> doHit) throws Exception {
+        if (doHit.isPresent())
+            return ResponseEntity.status(HttpStatus.OK).body(service.getCrewPostAndMembersAndHit(crewPostNo));
         else
-            return ResponseEntity.status(HttpStatus.OK).body(service.getCrewPostAndMembersWith(crewPostNo));
+            return ResponseEntity.status(HttpStatus.OK).body(service.getCrewPostAndMembers(crewPostNo));
     }
 
     @PutMapping(value = "/crew/post/{crewPostNo}")
     public ResponseEntity<?> updatePost(CrewPostDTO dtoIn,
-                                        @PathVariable(value = "crewPostNo") Long crewPostNo) throws Exception {
-        service.updateCrewPost(dtoIn, crewPostNo);
-        return ResponseEntity.status(HttpStatus.OK).body("크루 게시글 수정 완료");
+                                        @PathVariable(value = "crewPostNo") Long crewPostNo,
+                                        @RequestHeader String userId) {
+        if (authorityUtil.isYou(userId, service.getWriterId(crewPostNo))) {
+            service.updateCrewPost(dtoIn, crewPostNo);
+            return ResponseEntity.status(HttpStatus.OK).body("크루 게시글 수정 완료");
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Authentication Fail");
+        }
     }
 
     @DeleteMapping(value = "/crew/post/{crewPostNo}")
-    public ResponseEntity<?> deletePost(@PathVariable(value = "crewPostNo") Long crewPostNo) throws Exception {
-        service.deleteCrewPost(crewPostNo);
+    public ResponseEntity<?> deletePost(@PathVariable(value = "crewPostNo") Long crewPostNo,
+                                        @RequestHeader String userId) {
+        if (authorityUtil.isYou(userId, service.getWriterId(crewPostNo))) {
+            service.deleteCrewPost(crewPostNo);
+            return ResponseEntity.status(HttpStatus.OK).body("크루 게시글 삭제 완료");
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Authentication Fail");
+        }
+    }
+
+    @DeleteMapping(value = "/crew/post/member/{memberId}")
+    public ResponseEntity<?> deletePostByMember(@PathVariable(value = "memberId") String memberId) {
+        service.deleteCrewPostByMemberId(memberId);
         return ResponseEntity.status(HttpStatus.OK).body("크루 게시글 삭제 완료");
     }
 
-    @PostMapping(value = "/crew/post")
-    public ResponseEntity<CrewPostDTO> createPost(CrewPostDTO dtoIn, String membersWith) throws Exception {
-        service.saveCrewPost(dtoIn, membersWith);
-        return ResponseEntity.status(HttpStatus.OK).body(null);
+    @DeleteMapping(value = "/crew/post/all/{crewNo}")
+    public ResponseEntity<?> deletePostByCrewNo(@PathVariable(value = "crewNo") Long crewNo) {
+        service.deleteCrewPostByCrewNo(crewNo);
+        return ResponseEntity.status(HttpStatus.OK).body("크루 게시글 삭제 완료");
     }
 
     @PostMapping(value = "/crew/post/images/{imageUUID}")
@@ -72,10 +97,8 @@ public class CrewPostController {
     @ResponseBody
     @GetMapping(value = "/crew/post/images/{imageUUID}/{imageFileName}",
             produces = {MediaType.IMAGE_PNG_VALUE, MediaType.IMAGE_JPEG_VALUE, MediaType.IMAGE_GIF_VALUE})
-    public byte[] getCrewImage(@PathVariable("imageUUID") String uuid,
-                               @PathVariable("imageFileName") String imageFileName,
-                               HttpServletRequest request) throws IOException, FileSizeLimitExceededException {
-
+    public byte[] getCrewPostImage(@PathVariable("imageUUID") String uuid,
+                                   @PathVariable("imageFileName") String imageFileName) throws IOException {
         String imgPath = System.getProperty("user.dir") + "/src/main/resources/static/crew_post_image/" + uuid + "/" + imageFileName;
         InputStream in = new FileInputStream(imgPath);
         byte[] imageByteArr = IOUtils.toByteArray(in);
@@ -84,16 +107,26 @@ public class CrewPostController {
     }
 
     @PostMapping(value = "/crew/post/{crewPostNo}/member")
-    public ResponseEntity<?> addMemberToPost(@PathVariable(value = "crewPostNo") Long crewPostNo,
-                                             @RequestParam String memberId) {
-        service.saveMembers(crewPostNo, memberId);
-        return ResponseEntity.status(HttpStatus.OK).body("member added");
+    public ResponseEntity<?> tagMembers(@PathVariable(value = "crewPostNo") Long crewPostNo,
+                                        @RequestParam String memberId,
+                                        @RequestHeader String userId) {
+        if (authorityUtil.isYou(userId, service.getWriterId(crewPostNo))) {
+            service.saveMembers(crewPostNo, memberId);
+            return ResponseEntity.status(HttpStatus.OK).body("member added");
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Authentication Fail");
+        }
     }
 
     @DeleteMapping(value = "/crew/post/{crewPostNo}/member")
-    public ResponseEntity<?> deleteMemberInPost(@PathVariable(value = "crewPostNo") Long crewPostNo,
-                                                @RequestParam(value = "membersWithNo") Long membersWithNo) {
-        service.deleteMemberInPost(membersWithNo);
-        return ResponseEntity.status(HttpStatus.OK).body("member deleted");
+    public ResponseEntity<?> untagMembers(@PathVariable(value = "crewPostNo") Long crewPostNo,
+                                          @RequestParam String memberId,
+                                          @RequestHeader String userId) {
+        if (authorityUtil.isYou(userId, memberId)) {
+            service.deleteTaggedMember(new TaggedMemberId(crewPostNo, memberId));
+            return ResponseEntity.status(HttpStatus.OK).body("member deleted");
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Authentication Fail");
+        }
     }
 }
